requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
lxml==4.9.1
peewee==3.15.2
psycopg2-binary==2.9.3
python-decouple==3.6
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
